Rename class VerifiedSetOptions to SetOptions
